requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary
jinja2==3.1.3
python-multipart==0.0.9
starlette==0.35.0
pydantic>=2.0
